Move src files out of utils into src
