+++ src/emulator.js
