+++ src/flow-token.js
@@ -20,7 +20,7 @@ import * as types from "@onflow/types";
 import { defaultsByName } from "./file";
 import { replaceImportAddresses } from "./imports";
 import { executeScript, sendTransaction } from "./interaction";
-import { makeGetBalance, makeMintTransaction } from "../templates";
+import { makeGetBalance, makeMintTransaction } from "./templates";
 
 /**
  * Returns current FlowToken balance of account specified by address
