+++ src/interaction.js
