+++ src/manager.js
